# Conflicts:
#	Content.Server/Shuttles/Systems/EmergencyShuttleSystem.Console.cs
#	Content.Server/Shuttles/Systems/ShuttleConsoleSystem.cs
#	Content.Server/Shuttles/Systems/ShuttleSystem.EmergencyShuttle.cs

+++ Pow3r/Program.Veldrid.cs
@@ -241,7 +241,7 @@ void main()
 
             _vdSetTexture.Name = "SetTexture";
 
-            io.Fonts.SetTexID((nint) 0);
+            io.Fonts.SetTexID(0);
             io.Fonts.ClearTexData();
 
             _vdGfxDevice.ResizeMainWindow((uint) w, (uint) h);
